+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.6.5] - 2018-12-04 15:31:00
+### Fixed
+- Optimize product import
+### Added
+- Smart navigation in native category url support
+
 ## [4.6.3] - 2018-11-05 15:31:00
 ### Fixed
 - Optimize product import